partial commit from c9b0b62a5a9 split whatif sdk client
