+++ src/Resources/ResourceManager/Implementation/DeploymentStacks/NewAzManagementGroupDeploymentStack.cs
@@ -1,4 +1,4 @@
-﻿// ----------------------------------------------------------------------------------
+// ----------------------------------------------------------------------------------
 //
 // Copyright Microsoft Corporation
 // Licensed under the Apache License, Version 2.0 (the "License");
@@ -113,7 +113,7 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.Implementation
                 // Handle WhatIf scenario
                 if (MyInvocation.BoundParameters.ContainsKey("WhatIf") && ((SwitchParameter)MyInvocation.BoundParameters["WhatIf"]).ToBool())
                 {
-                    var whatIfResult = DeploymentStacksSdkClient.ExecuteManagementGroupDeploymentStackWhatIf(
+                    var whatIfResult = DeploymentStacksWhatIfSdkClient.ExecuteManagementGroupDeploymentStackWhatIf(
                         deploymentStackName: Name,
                         managementGroupId: ManagementGroupId,
                         location: Location,
