+++ src/Resources/ResourceManager/Implementation/DeploymentStacksWhatIf/NewAzManagementGroupDeploymentStackWhatIf.cs
