+++ src/Resources/ResourceManager/Implementation/DeploymentStacksWhatIf/NewAzSubscriptionDeploymentStackWhatIf.cs
