+++ src/Resources/ResourceManager/Implementation/DeploymentStacksWhatIf/RemoveAzResourceGroupDeploymentStackWhatIf.cs
@@ -19,29 +19,31 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.Implementation.Deploy
     using Microsoft.Azure.Commands.ResourceManager.Cmdlets.Components;
     using Microsoft.Azure.Commands.ResourceManager.Cmdlets.SdkModels.Deployments;
     using Microsoft.Azure.Commands.ResourceManager.Common;
+    using Microsoft.Azure.Commands.ResourceManager.Common.ArgumentCompleters;
 
     /// <summary>
-    /// Deletes an existing WhatIf result for a Management Group Deployment Stack.
+    /// Deletes an existing WhatIf result for a Resource Group Deployment Stack.
     /// </summary>
-    [Cmdlet("Remove", AzureRMConstants.AzureRMPrefix + "ManagementGroupDeploymentStackWhatIfResult",
-        SupportsShouldProcess = true, DefaultParameterSetName = RemoveByNameAndManagementGroupIdParameterSetName), OutputType(typeof(bool))]
-    public class RemoveAzManagementGroupDeploymentStackWhatIf : DeploymentStacksCmdletBase
+    [Cmdlet("Remove", AzureRMConstants.AzureRMPrefix + "ResourceGroupDeploymentStackWhatIfResult",
+        SupportsShouldProcess = true, DefaultParameterSetName = RemoveByNameAndResourceGroupNameParameterSetName), OutputType(typeof(bool))]
+    public class RemoveAzResourceGroupDeploymentStackWhatIf : DeploymentStacksCmdletBase
     {
         internal const string RemoveByResourceIdParameterSetName = "RemoveByResourceId";
-        internal const string RemoveByNameAndManagementGroupIdParameterSetName = "RemoveByNameAndManagementGroupId";
+        internal const string RemoveByNameAndResourceGroupNameParameterSetName = "RemoveByNameAndResourceGroupName";
         internal const string RemoveByInputObjectParameterSetName = "RemoveByInputObject";
 
         [Parameter(Position = 1, Mandatory = true, ValueFromPipelineByPropertyName = true,
-            ParameterSetName = RemoveByNameAndManagementGroupIdParameterSetName,
+            ParameterSetName = RemoveByNameAndResourceGroupNameParameterSetName,
             HelpMessage = "The name of the WhatIf result to delete.")]
         [ValidateNotNullOrEmpty]
         public string Name { get; set; }
 
         [Parameter(Position = 0, Mandatory = true, ValueFromPipelineByPropertyName = true,
-            ParameterSetName = RemoveByNameAndManagementGroupIdParameterSetName,
-            HelpMessage = "The ID of the management group.")]
+            ParameterSetName = RemoveByNameAndResourceGroupNameParameterSetName,
+            HelpMessage = "The name of the ResourceGroup containing the WhatIf result.")]
+        [ResourceGroupCompleter]
         [ValidateNotNullOrEmpty]
-        public string ManagementGroupId { get; set; }
+        public string ResourceGroupName { get; set; }
 
         [Alias("Id")]
         [Parameter(Mandatory = true, ValueFromPipelineByPropertyName = true,
@@ -71,36 +73,24 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.Implementation.Deploy
                     ResourceId = InputObject.Id;
                 }
 
-                // Resolve Name and ManagementGroupId if ResourceId was provided
-                if (!string.IsNullOrEmpty(ResourceId))
-                {
-                    Name = Name ?? ResourceIdUtility.GetDeploymentName(ResourceId);
-                    if (string.IsNullOrEmpty(ManagementGroupId))
-                    {
-                        // Extract management group id from: /providers/Microsoft.Management/managementGroups/{mgId}/...
-                        var segments = ResourceId.Split('/');
-                        var mgIndex = Array.IndexOf(segments, "managementGroups");
-                        if (mgIndex >= 0 && mgIndex + 1 < segments.Length)
-                        {
-                            ManagementGroupId = segments[mgIndex + 1];
-                        }
-                    }
-                }
+                // Resolve Name and ResourceGroupName if ResourceId was provided
+                ResourceGroupName = ResourceGroupName ?? ResourceIdUtility.GetResourceGroupName(ResourceId);
+                Name = Name ?? ResourceIdUtility.GetDeploymentName(ResourceId);
 
-                if (Name == null || ManagementGroupId == null)
+                if (Name == null || ResourceGroupName == null)
                 {
                     throw new PSArgumentException($"Provided Id '{ResourceId}' is not in correct form. Should be in form " +
-                        "/providers/Microsoft.Management/managementGroups/<mgId>/providers/Microsoft.Resources/deploymentStacksWhatIfResults/<name>");
+                        "/subscriptions/<subid>/resourceGroups/<rgname>/providers/Microsoft.Resources/deploymentStacksWhatIfResults/<name>");
                 }
 
                 ConfirmAction(
                     Force.IsPresent,
-                    $"Are you sure you want to delete WhatIf result '{Name}' in ManagementGroup '{ManagementGroupId}'?",
+                    $"Are you sure you want to delete WhatIf result '{Name}' in ResourceGroup '{ResourceGroupName}'?",
                     "Delete",
                     Name,
                     () =>
                     {
-                        DeploymentStacksSdkClient.DeleteManagementGroupDeploymentStackWhatIfResult(ManagementGroupId, Name);
+                        DeploymentStacksWhatIfSdkClient.DeleteResourceGroupDeploymentStackWhatIfResult(ResourceGroupName, Name);
                         if (PassThru.IsPresent)
                         {
                             WriteObject(true);
