+++ src/Resources/ResourceManager/Implementation/DeploymentStacksWhatIf/SetAzSubscriptionDeploymentStackWhatIf.cs
