+++ src/Resources/ResourceManager/SdkClient/DeploymentStacksSdkClient.cs
@@ -1178,565 +1178,5 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.SdkClient
 
             return convertedErrors;
         }
-
-        private PSDeploymentStackWhatIfResult ConvertToPSDeploymentStackWhatIfResult(object sdkWhatIfResult)
-        {
-            var json = JsonConvert.SerializeObject(sdkWhatIfResult);
-            return JsonConvert.DeserializeObject<PSDeploymentStackWhatIfResult>(json);
-        }
-
-        public PSDeploymentStackWhatIfResult GetResourceGroupDeploymentStackWhatIfResult(string resourceGroupName, string stackName, bool withPropertyChanges = false, bool throwIfNotExists = true)
-        {
-            try
-            {
-                var result = withPropertyChanges
-                    ? DeploymentStacksClient.DeploymentStacksWhatIfResultsAtResourceGroup.WhatIf(resourceGroupName, stackName)
-                    : DeploymentStacksClient.DeploymentStacksWhatIfResultsAtResourceGroup.Get(resourceGroupName, stackName);
-                return ConvertToPSDeploymentStackWhatIfResult(result);
-            }
-            catch (Exception ex)
-            {
-                if (ex is ErrorResponseException dex)
-                {
-                    if (dex.Response.StatusCode == System.Net.HttpStatusCode.NotFound && !throwIfNotExists)
-                    {
-                        return null;
-                    }
-
-                    throw new PSArgumentException(dex.Body?.Error?.Message ?? dex.Message);
-                }
-                throw;
-            }
-        }
-
-        public PSDeploymentStackWhatIfResult GetResourceGroupDeploymentStackWhatIfResultWithPropertyChanges(string resourceGroupName, string stackName)
-        {
-            return GetResourceGroupDeploymentStackWhatIfResult(resourceGroupName, stackName, withPropertyChanges: true);
-        }
-
-        public IList<PSDeploymentStackWhatIfResult> ListResourceGroupDeploymentStackWhatIfResults(string resourceGroupName)
-        {
-            try
-            {
-                var list = new List<PSDeploymentStackWhatIfResult>();
-                var results = DeploymentStacksClient.DeploymentStacksWhatIfResultsAtResourceGroup.List(resourceGroupName);
-                list.AddRange(results.Select(r => ConvertToPSDeploymentStackWhatIfResult(r)));
-                while (results.NextPageLink != null)
-                {
-                    results = DeploymentStacksClient.DeploymentStacksWhatIfResultsAtResourceGroup.ListNext(results.NextPageLink);
-                    list.AddRange(results.Select(r => ConvertToPSDeploymentStackWhatIfResult(r)));
-                }
-
-                return list;
-            }
-            catch (Exception ex)
-            {
-                if (ex is ErrorResponseException dex)
-                    throw new PSArgumentException(dex.Body?.Error?.Message ?? dex.Message);
-                throw;
-            }
-        }
-
-        public void DeleteResourceGroupDeploymentStackWhatIfResult(string resourceGroupName, string stackName)
-        {
-            try
-            {
-                DeploymentStacksClient.DeploymentStacksWhatIfResultsAtResourceGroup.Delete(resourceGroupName, stackName);
-            }
-            catch (Exception ex)
-            {
-                if (ex is ErrorResponseException dex)
-                    throw new PSArgumentException(dex.Body?.Error?.Message ?? dex.Message);
-                throw;
-            }
-        }
-
-        public PSDeploymentStackWhatIfResult GetSubscriptionDeploymentStackWhatIfResult(string stackName, bool withPropertyChanges = false, bool throwIfNotExists = true)
-        {
-            try
-            {
-                var result = withPropertyChanges
-                    ? DeploymentStacksClient.DeploymentStacksWhatIfResultsAtSubscription.WhatIf(stackName)
-                    : DeploymentStacksClient.DeploymentStacksWhatIfResultsAtSubscription.Get(stackName);
-                return ConvertToPSDeploymentStackWhatIfResult(result);
-            }
-            catch (Exception ex)
-            {
-                if (ex is ErrorResponseException dex)
-                {
-                    if (dex.Response.StatusCode == System.Net.HttpStatusCode.NotFound && !throwIfNotExists)
-                    {
-                        return null;
-                    }
-
-                    throw new PSArgumentException(dex.Body?.Error?.Message ?? dex.Message);
-                }
-                throw;
-            }
-        }
-
-        public PSDeploymentStackWhatIfResult GetSubscriptionDeploymentStackWhatIfResultWithPropertyChanges(string stackName)
-        {
-            return GetSubscriptionDeploymentStackWhatIfResult(stackName, withPropertyChanges: true);
-        }
-
-        public IList<PSDeploymentStackWhatIfResult> ListSubscriptionDeploymentStackWhatIfResults()
-        {
-            try
-            {
-                var list = new List<PSDeploymentStackWhatIfResult>();
-                var results = DeploymentStacksClient.DeploymentStacksWhatIfResultsAtSubscription.List();
-                list.AddRange(results.Select(r => ConvertToPSDeploymentStackWhatIfResult(r)));
-                while (results.NextPageLink != null)
-                {
-                    results = DeploymentStacksClient.DeploymentStacksWhatIfResultsAtSubscription.ListNext(results.NextPageLink);
-                    list.AddRange(results.Select(r => ConvertToPSDeploymentStackWhatIfResult(r)));
-                }
-
-                return list;
-            }
-            catch (Exception ex)
-            {
-                if (ex is ErrorResponseException dex)
-                    throw new PSArgumentException(dex.Body?.Error?.Message ?? dex.Message);
-                throw;
-            }
-        }
-
-        public void DeleteSubscriptionDeploymentStackWhatIfResult(string stackName)
-        {
-            try
-            {
-                DeploymentStacksClient.DeploymentStacksWhatIfResultsAtSubscription.Delete(stackName);
-            }
-            catch (Exception ex)
-            {
-                if (ex is ErrorResponseException dex)
-                    throw new PSArgumentException(dex.Body?.Error?.Message ?? dex.Message);
-                throw;
-            }
-        }
-
-        public PSDeploymentStackWhatIfResult GetManagementGroupDeploymentStackWhatIfResult(string managementGroupId, string stackName, bool withPropertyChanges = false, bool throwIfNotExists = true)
-        {
-            try
-            {
-                var result = withPropertyChanges
-                    ? DeploymentStacksClient.DeploymentStacksWhatIfResultsAtManagementGroup.WhatIf(managementGroupId, stackName)
-                    : DeploymentStacksClient.DeploymentStacksWhatIfResultsAtManagementGroup.Get(managementGroupId, stackName);
-                return ConvertToPSDeploymentStackWhatIfResult(result);
-            }
-            catch (Exception ex)
-            {
-                if (ex is ErrorResponseException dex)
-                {
-                    if (dex.Response.StatusCode == System.Net.HttpStatusCode.NotFound && !throwIfNotExists)
-                    {
-                        return null;
-                    }
-
-                    throw new PSArgumentException(dex.Body?.Error?.Message ?? dex.Message);
-                }
-                throw;
-            }
-        }
-
-        public PSDeploymentStackWhatIfResult GetManagementGroupDeploymentStackWhatIfResultWithPropertyChanges(string managementGroupId, string stackName)
-        {
-            return GetManagementGroupDeploymentStackWhatIfResult(managementGroupId, stackName, withPropertyChanges: true);
-        }
-
-        public IList<PSDeploymentStackWhatIfResult> ListManagementGroupDeploymentStackWhatIfResults(string managementGroupId)
-        {
-            try
-            {
-                var list = new List<PSDeploymentStackWhatIfResult>();
-                var results = DeploymentStacksClient.DeploymentStacksWhatIfResultsAtManagementGroup.List(managementGroupId);
-                list.AddRange(results.Select(r => ConvertToPSDeploymentStackWhatIfResult(r)));
-                while (results.NextPageLink != null)
-                {
-                    results = DeploymentStacksClient.DeploymentStacksWhatIfResultsAtManagementGroup.ListNext(results.NextPageLink);
-                    list.AddRange(results.Select(r => ConvertToPSDeploymentStackWhatIfResult(r)));
-                }
-
-                return list;
-            }
-            catch (Exception ex)
-            {
-                if (ex is ErrorResponseException dex)
-                    throw new PSArgumentException(dex.Body?.Error?.Message ?? dex.Message);
-                throw;
-            }
-        }
-
-        public void DeleteManagementGroupDeploymentStackWhatIfResult(string managementGroupId, string stackName)
-        {
-            try
-            {
-                DeploymentStacksClient.DeploymentStacksWhatIfResultsAtManagementGroup.Delete(managementGroupId, stackName);
-            }
-            catch (Exception ex)
-            {
-                if (ex is ErrorResponseException dex)
-                    throw new PSArgumentException(dex.Body?.Error?.Message ?? dex.Message);
-                throw;
-            }
-        }
-
-        private DeploymentStacksWhatIfResult CreateDeploymentStacksWhatIfResult(
-            string location,
-            string templateFile,
-            string templateUri,
-            string templateSpec,
-            Hashtable templateObject,
-            string parameterUri,
-            Hashtable parameters,
-            string description,
-            string resourcesCleanupAction,
-            string resourceGroupsCleanupAction,
-            string managementGroupsCleanupAction,
-            string deploymentScope,
-            string denySettingsMode,
-            string[] denySettingsExcludedPrincipals,
-            string[] denySettingsExcludedActions,
-            bool denySettingsApplyToChildScopes,
-            string deploymentStackName,
-            string resourceGroupName,
-            string managementGroupId)
-        {
-            var actionOnUnmanage = new Microsoft.Azure.Management.Resources.DeploymentStacks.Models.ActionOnUnmanage
-            {
-                Resources = resourcesCleanupAction,
-                ResourceGroups = resourceGroupsCleanupAction,
-                ManagementGroups = managementGroupsCleanupAction
-            };
-
-            var denySettings = new Microsoft.Azure.Management.Resources.DeploymentStacks.Models.DenySettings
-            {
-                Mode = denySettingsMode,
-                ExcludedPrincipals = denySettingsExcludedPrincipals,
-                ExcludedActions = denySettingsExcludedActions,
-                ApplyToChildScopes = denySettingsApplyToChildScopes
-            };
-
-            string stackResourceId;
-            if (!string.IsNullOrEmpty(resourceGroupName))
-            {
-                stackResourceId = $"/subscriptions/{azureContext.Subscription.Id}/resourceGroups/{resourceGroupName}/providers/Microsoft.Resources/deploymentStacks/{deploymentStackName}";
-            }
-            else if (!string.IsNullOrEmpty(managementGroupId))
-            {
-                stackResourceId = $"/providers/Microsoft.Management/managementGroups/{managementGroupId}/providers/Microsoft.Resources/deploymentStacks/{deploymentStackName}";
-            }
-            else
-            {
-                stackResourceId = $"/subscriptions/{azureContext.Subscription.Id}/providers/Microsoft.Resources/deploymentStacks/{deploymentStackName}";
-            }
-
-            var properties = new DeploymentStacksWhatIfResultProperties(
-                actionOnUnmanage: actionOnUnmanage,
-                denySettings: denySettings,
-                deploymentStackResourceId: stackResourceId,
-                retentionInterval: TimeSpan.FromHours(2),
-                deploymentScope: deploymentScope,
-                description: description);
-
-            if (templateSpec != null)
-            {
-                properties.TemplateLink = new DeploymentStacksTemplateLink
-                {
-                    Id = templateSpec,
-                };
-            }
-            else if (Uri.IsWellFormedUriString(templateUri, UriKind.Absolute))
-            {
-                properties.TemplateLink = new DeploymentStacksTemplateLink
-                {
-                    Uri = templateUri,
-                };
-            }
-            else if (!string.IsNullOrEmpty(templateFile))
-            {
-                properties.Template = FileUtilities.DataStore.ReadFileAsStream(templateFile).FromJson<JObject>().ToObject<Dictionary<string, object>>();
-            }
-            else
-            {
-                properties.Template = templateObject.ToJToken().ToObject<Dictionary<string, object>>();
-            }
-
-            if (Uri.IsWellFormedUriString(parameterUri, UriKind.Absolute))
-            {
-                properties.ParametersLink = new DeploymentStacksParametersLink
-                {
-                    Uri = parameterUri
-                };
-            }
-            else if (parameters != null)
-            {
-                properties.Parameters = ConvertParameterHashtableToDictionary(parameters);
-            }
-
-            return new DeploymentStacksWhatIfResult
-            {
-                Location = location,
-                Properties = properties
-            };
-        }
-
-        public PSDeploymentStackWhatIfResult ExecuteDeploymentStackWhatIf(PSDeploymentStackWhatIfParameters parameters)
-        {
-            if (parameters == null)
-            {
-                throw new ArgumentNullException(nameof(parameters));
-            }
-
-            if (!string.IsNullOrEmpty(parameters.ResourceGroupName))
-            {
-                return ExecuteResourceGroupDeploymentStackWhatIf(
-                    parameters.StackName,
-                    parameters.ResourceGroupName,
-                    parameters.TemplateFile,
-                    parameters.TemplateUri,
-                    parameters.TemplateSpecId,
-                    parameters.TemplateObject,
-                    parameters.TemplateParameterUri,
-                    parameters.TemplateParameterObject,
-                    parameters.Description,
-                    parameters.ResourcesCleanupAction,
-                    parameters.ResourceGroupsCleanupAction,
-                    parameters.ManagementGroupsCleanupAction,
-                    parameters.DenySettingsMode,
-                    parameters.DenySettingsExcludedPrincipals,
-                    parameters.DenySettingsExcludedActions,
-                    parameters.DenySettingsApplyToChildScopes,
-                    false);
-            }
-
-            if (!string.IsNullOrEmpty(parameters.ManagementGroupId))
-            {
-                return ExecuteManagementGroupDeploymentStackWhatIf(
-                    parameters.StackName,
-                    parameters.ManagementGroupId,
-                    parameters.Location,
-                    parameters.TemplateFile,
-                    parameters.TemplateUri,
-                    parameters.TemplateSpecId,
-                    parameters.TemplateObject,
-                    parameters.TemplateParameterUri,
-                    parameters.TemplateParameterObject,
-                    parameters.Description,
-                    parameters.ResourcesCleanupAction,
-                    parameters.ResourceGroupsCleanupAction,
-                    parameters.ManagementGroupsCleanupAction,
-                    parameters.DeploymentScope,
-                    parameters.DenySettingsMode,
-                    parameters.DenySettingsExcludedPrincipals,
-                    parameters.DenySettingsExcludedActions,
-                    parameters.DenySettingsApplyToChildScopes,
-                    false);
-            }
-
-            return ExecuteSubscriptionDeploymentStackWhatIf(
-                parameters.StackName,
-                parameters.Location,
-                parameters.TemplateFile,
-                parameters.TemplateUri,
-                parameters.TemplateSpecId,
-                parameters.TemplateObject,
-                parameters.TemplateParameterUri,
-                parameters.TemplateParameterObject,
-                parameters.Description,
-                parameters.ResourcesCleanupAction,
-                parameters.ResourceGroupsCleanupAction,
-                parameters.ManagementGroupsCleanupAction,
-                parameters.DeploymentScope,
-                parameters.DenySettingsMode,
-                parameters.DenySettingsExcludedPrincipals,
-                parameters.DenySettingsExcludedActions,
-                parameters.DenySettingsApplyToChildScopes,
-                false);
-        }
-
-        public PSDeploymentStackWhatIfResult ExecuteResourceGroupDeploymentStackWhatIf(
-            string deploymentStackName,
-            string resourceGroupName,
-            string templateFile,
-            string templateUri,
-            string templateSpec,
-            Hashtable templateObject,
-            string parameterUri,
-            Hashtable parameters,
-            string description,
-            string resourcesCleanupAction,
-            string resourceGroupsCleanupAction,
-            string managementGroupsCleanupAction,
-            string denySettingsMode,
-            string[] denySettingsExcludedPrincipals,
-            string[] denySettingsExcludedActions,
-            bool denySettingsApplyToChildScopes,
-            bool bypassStackOutOfSyncError)
-        {
-            try
-            {
-                WriteVerbose($"Executing What-If for deployment stack '{deploymentStackName}' in resource group '{resourceGroupName}'");
-
-                var whatIfRequest = CreateDeploymentStacksWhatIfResult(
-                    location: null,
-                    templateFile,
-                    templateUri,
-                    templateSpec,
-                    templateObject,
-                    parameterUri,
-                    parameters,
-                    description,
-                    resourcesCleanupAction,
-                    resourceGroupsCleanupAction,
-                    managementGroupsCleanupAction,
-                    deploymentScope: null,
-                    denySettingsMode,
-                    denySettingsExcludedPrincipals,
-                    denySettingsExcludedActions,
-                    denySettingsApplyToChildScopes,
-                    deploymentStackName,
-                    resourceGroupName,
-                    null);
-
-                var whatIfResult = DeploymentStacksClient.DeploymentStacksWhatIfResultsAtResourceGroup.CreateOrUpdate(
-                    resourceGroupName,
-                    deploymentStackName,
-                    whatIfRequest);
-
-                WriteVerbose("Retrieving What-If result with property changes...");
-                var postResult = DeploymentStacksClient.DeploymentStacksWhatIfResultsAtResourceGroup.WhatIf(resourceGroupName, deploymentStackName);
-                return ConvertToPSDeploymentStackWhatIfResult(postResult);
-            }
-            catch (Exception ex)
-            {
-                WriteError($"Error executing What-If: {ex.Message}");
-                throw;
-            }
-        }
-
-        public PSDeploymentStackWhatIfResult ExecuteSubscriptionDeploymentStackWhatIf(
-            string deploymentStackName,
-            string location,
-            string templateFile,
-            string templateUri,
-            string templateSpec,
-            Hashtable templateObject,
-            string parameterUri,
-            Hashtable parameters,
-            string description,
-            string resourcesCleanupAction,
-            string resourceGroupsCleanupAction,
-            string managementGroupsCleanupAction,
-            string deploymentScope,
-            string denySettingsMode,
-            string[] denySettingsExcludedPrincipals,
-            string[] denySettingsExcludedActions,
-            bool denySettingsApplyToChildScopes,
-            bool bypassStackOutOfSyncError)
-        {
-            try
-            {
-                WriteVerbose($"Executing What-If for deployment stack '{deploymentStackName}' at subscription scope");
-
-                var whatIfRequest = CreateDeploymentStacksWhatIfResult(
-                    location,
-                    templateFile,
-                    templateUri,
-                    templateSpec,
-                    templateObject,
-                    parameterUri,
-                    parameters,
-                    description,
-                    resourcesCleanupAction,
-                    resourceGroupsCleanupAction,
-                    managementGroupsCleanupAction,
-                    deploymentScope,
-                    denySettingsMode,
-                    denySettingsExcludedPrincipals,
-                    denySettingsExcludedActions,
-                    denySettingsApplyToChildScopes,
-                    deploymentStackName,
-                    null,
-                    null);
-
-                var whatIfResult = DeploymentStacksClient.DeploymentStacksWhatIfResultsAtSubscription.CreateOrUpdate(
-                    deploymentStackName,
-                    whatIfRequest);
-
-                WriteVerbose("Retrieving What-If result with property changes...");
-                var postResult = DeploymentStacksClient.DeploymentStacksWhatIfResultsAtSubscription.WhatIf(deploymentStackName);
-                return ConvertToPSDeploymentStackWhatIfResult(postResult);
-            }
-            catch (Exception ex)
-            {
-                WriteError($"Error executing What-If: {ex.Message}");
-                throw;
-            }
-        }
-
-        public PSDeploymentStackWhatIfResult ExecuteManagementGroupDeploymentStackWhatIf(
-            string deploymentStackName,
-            string managementGroupId,
-            string location,
-            string templateFile,
-            string templateUri,
-            string templateSpec,
-            Hashtable templateObject,
-            string parameterUri,
-            Hashtable parameters,
-            string description,
-            string resourcesCleanupAction,
-            string resourceGroupsCleanupAction,
-            string managementGroupsCleanupAction,
-            string deploymentScope,
-            string denySettingsMode,
-            string[] denySettingsExcludedPrincipals,
-            string[] denySettingsExcludedActions,
-            bool denySettingsApplyToChildScopes,
-            bool bypassStackOutOfSyncError)
-        {
-            try
-            {
-                WriteVerbose($"Executing What-If for deployment stack '{deploymentStackName}' in management group '{managementGroupId}'");
-
-                var whatIfRequest = CreateDeploymentStacksWhatIfResult(
-                    location,
-                    templateFile,
-                    templateUri,
-                    templateSpec,
-                    templateObject,
-                    parameterUri,
-                    parameters,
-                    description,
-                    resourcesCleanupAction,
-                    resourceGroupsCleanupAction,
-                    managementGroupsCleanupAction,
-                    deploymentScope,
-                    denySettingsMode,
-                    denySettingsExcludedPrincipals,
-                    denySettingsExcludedActions,
-                    denySettingsApplyToChildScopes,
-                    deploymentStackName,
-                    null,
-                    managementGroupId);
-
-                var whatIfResult = DeploymentStacksClient.DeploymentStacksWhatIfResultsAtManagementGroup.CreateOrUpdate(
-                    managementGroupId,
-                    deploymentStackName,
-                    whatIfRequest);
-
-                WriteVerbose("Retrieving What-If result with property changes...");
-                var postResult = DeploymentStacksClient.DeploymentStacksWhatIfResultsAtManagementGroup.WhatIf(managementGroupId, deploymentStackName);
-                return ConvertToPSDeploymentStackWhatIfResult(postResult);
-            }
-            catch (Exception ex)
-            {
-                WriteError($"Error executing What-If: {ex.Message}");
-                throw;
-            }
-        }
     }
 }
